requirements.txt
flask
transformers
gradio
gunicorn
python-dotenv
google-genai
torch
requests
